API

MOONS.Grids.DataLocation — Type

DataLocation

Subtypes are used for dispatching on the data location

MOONS.Grids.Center1D — Type

1D cell center

MOONS.Grids.Vertex1D — Type

1D cell vertex

MOONS.Grids.Coordinates — Type

Coordinates

Cell node (n) and center (c) 1D coordinates.

MOONS.Grids.ghost_dual — Function

ghost_dual(::Extremum)

A 2-element vector near the boundary with 1's on ghost cells and 0's on interior cells

MOONS.Grids.binary — Function

binary(::Extremum)

Returns 0 for Min and 1 for Max

MOONS.Grids.n_hat — Function

n_hat(::Extremum)

The outward normal vector to the boundary

MOONS.Grids.ghost_vec — Function

ghost_vec(::Extremum)

A 3-element vector near the boundary with 1's on ghost cells and 0's on interior cells

MOONS.Grids.Boundary — Type

Type for dispatching on boundary points

MOONS.Grids.Interior — Type

Type for dispatching on interior points

MOONS.Grids.All — Type

Type for dispatching on all points

MOONS.Grids.NodeBoundary — Type

Node boundary type, containing ghost, boundary, and first interior indices

MOONS.Grids.CellBoundary — Type

Cell center boundary type, containing ghost and first interior indices

MOONS.Grids.Ghost — Type

Type for dispatching on ghost points

MOONS.Grids.Grid — Type

Grid

An N-dimensional Cartesian, structured, computational grid.

MOONS.Grids.Max — Type

Max

A type for dispatching on the maximum of the domain along a particular direction

MOONS.Grids.Min — Type

Min

A type for dispatching on the minimum of the domain along a particular direction